+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Full changelog](https://github.com/alexcrichton/hermes/compare/v1.1.6...master)
 
 * [added] New UI for viewing song history in a drawer instead of a popup
+* [added] Stations can now be created by genre as well as by seed
 
 # Version 1.1.6 (6/17/12)